SLING-10288 - FelixFrameworkScanner sets framework properties always for current Java version
Reduce duplication when instantiating bundle descriptors for the system bundle.
